TD 2/app_cdc/requirements.txt
mysql-connector-python==8.0.33
psycopg2-binary==2.9.9
